move it_knows_if_it_has_only_some_tissue from unit to integration
